fix: integer/float

+++ src/models/Pokemon.js
@@ -27,10 +27,10 @@ module.exports = (sequelize) => {
       type: DataTypes.INTEGER
     },
     height: {
-      type: DataTypes.INTEGER
+      type: DataTypes.FLOAT
     },
     weight: {
-      type: DataTypes.INTEGER
+      type: DataTypes.FLOAT
     },
     image: {
       type: DataTypes.STRING
